requirements.txt
streamlit==1.37.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.5
easyocr==1.7.1
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
